% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_bibliography}
\alias{aggregate_bibliography}
\title{aggregate bibliography}
\usage{
aggregate_bibliography(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
aggregated bib file
}
\description{
aggregate bibliograhy to fill in the missing references
}
\examples{
test_folder <- system.file("aggr_example", package = "rebib")
rebib::aggregate_bibliography(test_folder)
readLines(paste(test_folder,"example.bib",sep="/"))
}
