% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{biblio_converter}
\alias{biblio_converter}
\title{bibliography converter}
\usage{
biblio_converter(file_path = "", log_rebib = FALSE)
}
\arguments{
\item{file_path}{provide a file_path with file name to point tex/bbl file}

\item{log_rebib}{option to enable log files for rebib}
}
\value{
bib file
}
\description{
a quick converter for bbl/tex to bib
}
\examples{
test_file <- system.file("standalone/sample.bbl", package = "rebib")
rebib::biblio_converter(file_path = test_file)
head(readLines(xfun::with_ext(test_file,"bib")))
}
