% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{biblio_convertor}
\alias{biblio_convertor}
\title{biblio convertor}
\usage{
biblio_convertor(file_path = "")
}
\arguments{
\item{file_path}{provide a file_path with file name to point tex/bbl file}
}
\value{
bib file
}
\description{
a quick convertor for bbl/tex to bib
}
