% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bibliography_exists}
\alias{bibliography_exists}
\title{bibliography exists}
\usage{
bibliography_exists(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
TRUE/FALSE
}
\description{
check if embedded bibliography exists in the latex file or not
}
\examples{
wd <-  system.file("article", package = "rebib")
# Only reads the article file
rebib::bibliography_exists(wd)
}
