% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{bibliography_parser}
\alias{bibliography_parser}
\title{Main parser logic for converting inbuilt latex bibliography}
\usage{
bibliography_parser(single_bib_data)
}
\arguments{
\item{single_bib_data}{}
}
\value{
bib_record with unique_id, author and
        title(also containing other info)
}
\description{
A function meant to be used internally or in a sequence as described
in the documentation
}
