% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{bibtex_reader}
\alias{bibtex_reader}
\title{bibtex reader}
\usage{
bibtex_reader(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
list of BibTex references
}
\description{
bibtex reader
}
