% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bibtex_writer}
\alias{bibtex_writer}
\title{bibtex writer
writes bibtex data in a structured format to the .bib file}
\usage{
bibtex_writer(bibtex_data, file_name)
}
\arguments{
\item{bibtex_data}{a list of minimal bibtex data}

\item{file_name}{name of the tex file}
}
\value{
a bibtex file is written
}
\description{
bibtex writer
writes bibtex data in a structured format to the .bib file
}
