% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{citation_reader}
\alias{citation_reader}
\title{citation reader}
\usage{
citation_reader(file_path)
}
\arguments{
\item{file_path}{path to the LaTeX file}
}
\value{
count of the inline element
}
\description{
counts/reads  Cite inline elements embedded within the latex
file currently supported inlines \\cite,\\citealp, \\citep, \\citet
}
\examples{
file_path <- system.file("examples/article/example.tex",
                 package = "rebib")
cite <- rebib::citation_reader(file_path)
cite
}
