% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{count_inline}
\alias{count_inline}
\title{count citation elements}
\usage{
count_inline(file_path)
}
\arguments{
\item{file_path}{path to the LaTeX file}
}
\value{
count of the inline element
}
\description{
counts Cite inline elements embedded within the latex file
currently supported inlines \\cite,\\citealp, \\citep, \\citet
}
\examples{
file_path <- system.file("examples/article",
                 package = "rebib")
cite <- rebib::count_inline(file_path)
}
