% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extract_embeded_bib_items}
\alias{extract_embeded_bib_items}
\title{extract the bibliography in chunks seperated at bibitem}
\usage{
extract_embeded_bib_items(article_dir = "", file_name = "", file_path = "")
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{file_name}{name of the file}

\item{file_path}{absolute path of file with file_name}
}
\value{
a list of bib entries separated at bibitem
}
\description{
intended to be an internal function which is used with other
 functions in flow.
}
