% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{filter_bbl_data}
\alias{filter_bbl_data}
\title{comment filter for bbl data}
\usage{
filter_bbl_data(bbl_data)
}
\arguments{
\item{bbl_data}{blocks of bib_data}
}
\value{
filtered bbl_data
}
\description{
removes commented latex lines to avoid wrong reference data
}
