% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citation.R
\name{filter_citation_references}
\alias{filter_citation_references}
\title{filter citation references}
\usage{
filter_citation_references(citation_references)
}
\arguments{
\item{citation_references}{a vector of citation references(unfiltered)}
}
\value{
a vector of filtered citation references
}
\description{
a filter for extracting citation references
}
