% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{filter_repetition}
\alias{filter_repetition}
\title{filter repeated entries}
\usage{
filter_repetition(bbl_data, bib_data)
}
\arguments{
\item{bbl_data}{parsed embedded bibliography data}

\item{bib_data}{bibtex reference data}
}
\value{
aggregated bbl_data
}
\description{
filter repeated entries
}
