% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_md_file_name}
\alias{get_md_file_name}
\title{get markdown file name}
\usage{
get_md_file_name(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
markdown file name (str)
}
\description{
finds the markdown file in directory which is referenced in the article
}
