% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{get_reference_name}
\alias{get_reference_name}
\title{get reference name}
\usage{
get_reference_name(bib_reference)
}
\arguments{
\item{bib_reference}{first line containing the cite reference}
}
\value{
reference name (str)
}
\description{
get reference name
}
\examples{
ref_first_line <- "@book{ihaka:1996,"
ref_name <- rebib::get_reference_name(ref_first_line)
ref_name
}
