% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{get_reference_type}
\alias{get_reference_type}
\title{get reference type}
\usage{
get_reference_type(bib_reference)
}
\arguments{
\item{bib_reference}{first line containing the cite reference}
}
\value{
reference type (str)
}
\description{
get reference type
}
\examples{
ref_first_line <- "@book{ihaka:1996,"
ref_type <- rebib::get_reference_type(ref_first_line)
ref_type
}
