% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_wrapper_type}
\alias{get_wrapper_type}
\title{get wrapper file name}
\usage{
get_wrapper_type(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
String with name of wrapper file or empty
}
\description{
This function gets the wrapper file name from the
commonly named R-Journal wrapper files.
}
\details{
Usually the R journal wrapper files are named either
1. RJwrapper.tex\\n
2. RJwrap.tex\\n
3. wrapper.tex\\n
}
