% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{log_setup}
\alias{log_setup}
\title{rebib log setup}
\usage{
log_setup(article_dir)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}
}
\value{
null
}
\description{
a wrapper function for logger package to set up log file for
logging
}
\examples{
 wd <-  system.file("article", package = "rebib")
rebib::log_setup(wd)
}
