% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibtex.R
\name{split_bibtex_references}
\alias{split_bibtex_references}
\title{split BibTex references}
\usage{
split_bibtex_references(bib_path)
}
\arguments{
\item{bib_path}{path to the bibtex file to be read}
}
\value{
list of references separated as types and names based on indices
}
\description{
split BibTex references
}
\examples{
wd <-  system.file("article", package = "rebib")
rebib::handle_bibliography(wd)
bib_path <- paste(wd, "example.bib", sep = "/")
references <- rebib::split_bibtex_references(bib_path)
print(references)
}
