% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_extract}
\alias{str_extract}
\title{extract substring}
\usage{
str_extract(x, patt)
}
\arguments{
\item{x}{object (Str)}

\item{patt}{pattern (Regex Str)}
}
\value{
list of extracted string elements
}
\description{
a wrapper for stringr::str_extract
}
