% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{str_split}
\alias{str_split}
\title{split string}
\usage{
str_split(x, patt)
}
\arguments{
\item{x}{object (Str)}

\item{patt}{pattern (Regex Str)}
}
\value{
list of vectorized string elements
}
\description{
a wrapper for stringr::str_split
}
