% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_bibliography}
\alias{aggregate_bibliography}
\title{aggregate bibliography}
\usage{
aggregate_bibliography(article_dir, log_rebib = FALSE)
}
\arguments{
\item{article_dir}{path to the directory which contains tex article}

\item{log_rebib}{option to enable log files for rebib}
}
\value{
aggregated bib file
}
\description{
aggregate bibliograhy to fill in the missing references
}
\examples{
dir.create(your_article_folder <- file.path(tempdir(), "exampledir"))
example_files <-  system.file("aggr_example", package = "rebib")
x <- file.copy(from = example_files,to=your_article_folder,recursive = TRUE)
your_article_path <- paste(your_article_folder,"aggr_example",sep="/")
rebib::aggregate_bibliography(your_article_path)
readLines(paste(your_article_path,"example.bib",sep="/"))
unlink(your_article_folder,recursive = TRUE)
}
