% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{biblio_convertor}
\alias{biblio_convertor}
\title{biblio convertor}
\usage{
biblio_convertor(file_path = "")
}
\arguments{
\item{file_path}{provide a file_path with file name to point tex/bbl file}
}
\value{
bib file
}
\description{
a quick convertor for bbl/tex to bib
}
\examples{
test_file <- system.file("standalone/sample.bbl", package = "rebib")
rebib::biblio_convertor(file_path = test_file)
head(readLines(xfun::with_ext(test_file,"bib")))
}
