% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{rebib_log}
\alias{rebib_log}
\title{log messages for various categories}
\usage{
rebib_log(message, category)
}
\arguments{
\item{message}{message to be sent}

\item{category}{category of the log message}
}
\value{
null
}
\description{
a wrapper function for logging different types of log entries
}
\examples{
 wd <-  system.file("article", package = "rebib")
rebib::log_setup(wd)
rebib::rebib_log("Hello", "INFO")
cat(readLines(paste(wd,"/rebib-log-",Sys.Date(),".log",sep="")),sep="\n")
}
