% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log-tools.R
\name{rebib_log}
\alias{rebib_log}
\title{log messages for various categories}
\usage{
rebib_log(message, category, idx)
}
\arguments{
\item{message}{message to be sent}

\item{category}{category of the log message}

\item{idx}{index of log level}
}
\value{
null
}
\description{
a wrapper function for logging different types of log entries
}
\examples{
dir.create(your_article_folder <- file.path(tempdir(), "exampledir"))
example_files <-  system.file("article", package = "rebib")
x <- file.copy(from = example_files,to=your_article_folder,recursive = TRUE)
your_article_path <- paste(your_article_folder,"article",sep="/")
rebib::log_setup(your_article_path, "log-file.log", 2)
rebib::rebib_log("Hello", "INFO", 2)
cat(readLines(paste(your_article_path,"/log-file.log",sep="")),sep="\n")
unlink(your_article_folder,recursive = TRUE)
}
